requirements.txt
pytest==5.4.1
py-solc-x==0.8.1
web3>=5.7.0
eth-brownie>=1.6.4,<2.0.0